refactor: move from 128 to 96 for demand (#212)
* remove supply rollover

* add back in toEffectiveAmount for now

* fixup

* fix

* first step, find and replace ValueX7X7

* compiles

* refactor: passes most

* save

* passes most

* clean up tests

* update tests

* todo comments

* move currencyRaised* values to 128.128 form

* feat: remove restrictions on bid amount and price (#211)

* Remove bid min amount and bid max price

* Add MAX_AMOUNT to ConstantsLib

* clean up tests

* fix test

* fix gas tests

* find and replace 128 with 96

* fmt and gas

* remove MAX_AMOUNT

* save

* Fix bounds, lower fuzz runs

* Add Bid Max Price

* add natspec

* chore: add tests for calculateFill and natspec

* forge fmt

* Update readme

* Feedback and missing test for BID_MAX_PRICE

* Add extra invariant

* Add metrics and fix bid max price submission

* fix tests

* bump invariant depth to 1000

* forge fmt

* nits

* feedback

* fix gas

* fix poc test

+++ src/interfaces/ITokenCurrencyStorage.sol
@@ -47,7 +47,7 @@ interface ITokenCurrencyStorage {
     function token() external view returns (IERC20Minimal);
 
     /// @notice The total supply of tokens to sell
-    function totalSupply() external view returns (uint256);
+    function totalSupply() external view returns (uint128);
 
     /// @notice The recipient of any unsold tokens at the end of the auction
     function tokensRecipient() external view returns (address);
